*total LocationService refactor completed, testing phase

+++ app/src/main/java/com/kumpello/whereiseveryone/main/friends/model/Location.kt
@@ -1,3 +1,3 @@
 package com.kumpello.whereiseveryone.main.friends.model
 
-data class Location(val lat: Double, val lon: Double, val bearing: Double, val alt: Double)
+data class Location(val lat: Double, val lon: Double, val bearing: Float, val alt: Double)
